+++ src/main/java/com/wenxin2/marioverse/MarioverseCreativeTabs.java
@@ -154,6 +154,7 @@ public class MarioverseCreativeTabs {
             add(event, ItemRegistry.PLUMBER_POTTERY_SHERD);
 
             add(event, ItemRegistry.SUPER_MUSHROOM_SPAWN_EGG);
+            add(event, ItemRegistry.DASH_MUSHROOM_SPAWN_EGG);
             add(event, ItemRegistry.ONE_UP_MUSHROOM_SPAWN_EGG);
             add(event, ItemRegistry.MINI_MUSHROOM_SPAWN_EGG);
             add(event, ItemRegistry.MEGA_MUSHROOM_SPAWN_EGG);
@@ -745,7 +746,8 @@ public class MarioverseCreativeTabs {
                 ItemStack qwilfish = customName(ItemRegistry.PORCUPUFFER_SPAWN_EGG.get(), Component.literal("Qwilfish"));
 
                 addAfter(event, Items.ZOMBIFIED_PIGLIN_SPAWN_EGG, ItemRegistry.SUPER_MUSHROOM_SPAWN_EGG);
-                addAfter(event, ItemRegistry.SUPER_MUSHROOM_SPAWN_EGG, ItemRegistry.ONE_UP_MUSHROOM_SPAWN_EGG);
+                addAfter(event, ItemRegistry.SUPER_MUSHROOM_SPAWN_EGG, ItemRegistry.DASH_MUSHROOM_SPAWN_EGG);
+                addAfter(event, ItemRegistry.DASH_MUSHROOM_SPAWN_EGG, ItemRegistry.ONE_UP_MUSHROOM_SPAWN_EGG);
                 addAfter(event, ItemRegistry.ONE_UP_MUSHROOM_SPAWN_EGG, ItemRegistry.MINI_MUSHROOM_SPAWN_EGG);
                 addAfter(event, ItemRegistry.MINI_MUSHROOM_SPAWN_EGG, ItemRegistry.MEGA_MUSHROOM_SPAWN_EGG);
                 addAfter(event, ItemRegistry.MEGA_MUSHROOM_SPAWN_EGG, ItemRegistry.FIRE_FLOWER_SPAWN_EGG);