+++ src/main/java/com/wenxin2/marioverse/entities/power_ups/MiniMushroomEntity.java
@@ -22,7 +22,7 @@ import software.bernie.geckolib.constant.DefaultAnimations;
 import software.bernie.geckolib.util.GeckoLibUtil;
 
 public class MiniMushroomEntity extends MushroomEntity implements GeoEntity, PowerUpSource {
-    protected static final RawAnimation WALK = RawAnimation.begin().thenLoop("animation.mini_mushroom.walk");
+    protected static final RawAnimation WALK = RawAnimation.begin().thenLoop("move.walk");
     private final AnimatableInstanceCache geoCache = GeckoLibUtil.createInstanceCache(this);
 
     public MiniMushroomEntity(EntityType<? extends MiniMushroomEntity> entityType, Level world) {