+++ src/main/java/com/wenxin2/marioverse/entities/power_ups/MushroomEntity.java
@@ -17,7 +17,7 @@ import software.bernie.geckolib.constant.DefaultAnimations;
 import software.bernie.geckolib.util.GeckoLibUtil;
 
 public class MushroomEntity extends BaseMushroomEntity implements GeoEntity {
-    protected static final RawAnimation WALK_ANIM = RawAnimation.begin().thenLoop("animation.super_mushroom.walk");
+    protected static final RawAnimation WALK = RawAnimation.begin().thenLoop("move.walk");
     private final AnimatableInstanceCache cache = GeckoLibUtil.createInstanceCache(this);
 
     public MushroomEntity(EntityType<? extends MushroomEntity> entityType, Level world) {
@@ -43,7 +43,7 @@ public class MushroomEntity extends BaseMushroomEntity implements GeoEntity {
 
     protected <E extends GeoAnimatable> PlayState walkAnimation(final AnimationState<E> event) {
         if (event.isMoving())
-            return event.setAndContinue(WALK_ANIM);
+            return event.setAndContinue(WALK);
         return PlayState.STOP;
     }
 }